+++ ddth-akka-core/src/main/java/com/github/ddth/akka/cluster/MasterActor.java
@@ -4,6 +4,13 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.github.ddth.akka.AkkaUtils;
+import com.github.ddth.akka.cluster.messages.GetLeaderMessage;
+import com.github.ddth.akka.cluster.messages.GetLeaderResponseMessage;
+import com.github.ddth.akka.cluster.messages.GetNodesMessage;
+import com.github.ddth.akka.cluster.messages.GetNodesResponseMessage;
+import com.github.ddth.akka.cluster.messages.IsLeaderMessage;
+import com.github.ddth.akka.cluster.messages.IsLeaderResponseMessage;
+import com.github.ddth.akka.cluster.messages.RefreshClusterMembersMessage;
 
 import akka.actor.ActorRef;
 import akka.actor.ActorSystem;
@@ -13,6 +20,9 @@ import akka.cluster.ClusterEvent.MemberEvent;
 import akka.cluster.ClusterEvent.MemberRemoved;
 import akka.cluster.ClusterEvent.MemberUp;
 import akka.cluster.ClusterEvent.UnreachableMember;
+import akka.cluster.Member;
+import akka.cluster.MemberStatus;
+import scala.collection.JavaConverters;
 
 /**
  * Actor that keeps track of nodes within the cluster.
@@ -48,17 +58,63 @@ public class MasterActor extends BaseClusterActor {
     protected void initActor() throws Exception {
         super.initActor();
 
+        setHandleMessageAsync(false);
+
         // subscribe to cluster changes
         getCluster().subscribe(self(), ClusterEvent.initialStateAsEvents(), MemberEvent.class,
                 UnreachableMember.class);
 
-        // setup message handler
+        // setup message handlers
         addMessageHandler(ClusterEvent.MemberUp.class, this::eventMemberUp);
         addMessageHandler(ClusterEvent.MemberRemoved.class, this::eventMemberRemoved);
         addMessageHandler(ClusterEvent.UnreachableMember.class, (msg) -> {
             LOGGER.warn("Node [" + msg.member().address().toString() + "] with roles "
                     + msg.member().getRoles() + " detected as unreachable.");
         });
+
+        addMessageHandler(RefreshClusterMembersMessage.class, msg -> {
+            ClusterMemberUtils.resetNodes();
+            JavaConverters.asJavaIterable(getCluster().state().members()).forEach(m -> {
+                LOGGER.warn("" + m);
+                if (m.status() == MemberStatus.up()) {
+                    ClusterMemberUtils.addNode(m);
+                }
+            });
+        });
+        addMessageHandler(GetLeaderMessage.class, msg -> {
+            ActorRef sender = sender();
+            if (sender != null) {
+                Member member = ClusterMemberUtils.getLeader(msg.role);
+                sender.tell(new GetLeaderResponseMessage(msg.getId(), msg.role, member), self());
+            } else {
+                LOGGER.warn("Received message [" + GetLeaderMessage.class.getSimpleName()
+                        + "] but sender is null!");
+            }
+        });
+        addMessageHandler(IsLeaderMessage.class, msg -> {
+            ActorRef sender = sender();
+            if (sender != null) {
+                boolean result = ClusterMemberUtils.isLeader(msg.role, msg.node);
+                sender.tell(new IsLeaderResponseMessage(msg.getId(), msg.role, msg.node, result),
+                        self());
+            } else {
+                LOGGER.warn("Received message [" + IsLeaderMessage.class.getSimpleName()
+                        + "] but sender is null!");
+            }
+        });
+        addMessageHandler(GetNodesMessage.class, msg -> {
+            ActorRef sender = sender();
+            if (sender != null) {
+                sender.tell(new GetNodesResponseMessage(msg.getId(), msg.role,
+                        ClusterMemberUtils.getNodes(msg.role)), self());
+            } else {
+                LOGGER.warn("Received message [" + GetLeaderMessage.class.getSimpleName()
+                        + "] but sender is null!");
+            }
+        });
+
+        // initialize Cluster member state
+        self().tell(new RefreshClusterMembersMessage(), self());
     }
 
     /**