+++ ddth-akka-core/src/main/java/com/github/ddth/akka/cluster/messages/GetNodesMessage.java
@@ -1,0 +1,31 @@
+package com.github.ddth.akka.cluster.messages;
+
+import org.apache.commons.lang3.builder.ToStringBuilder;
+import org.apache.commons.lang3.builder.ToStringStyle;
+
+import com.github.ddth.akka.BaseMessage;
+import com.github.ddth.akka.cluster.MasterActor;
+
+/**
+ * Send this message to {@link MasterActor} to ask for set of nodes for a
+ * specific role.
+ * 
+ * @author Thanh Nguyen <btnguyen2k@gmail.com>
+ * @since 0.1.4
+ */
+public class GetNodesMessage extends BaseMessage {
+    private static final long serialVersionUID = "0.1.4".hashCode();
+
+    public final String role;
+
+    public GetNodesMessage(String role) {
+        this.role = role;
+    }
+
+    @Override
+    public String toString() {
+        ToStringBuilder tsb = new ToStringBuilder(this, ToStringStyle.SHORT_PREFIX_STYLE);
+        tsb.append("role", role).appendSuper(super.toString());
+        return tsb.toString();
+    }
+}